Extract WorkerJobs components folder out of Jobs folder
