+++ client/src/components/index.js
@@ -28,9 +28,9 @@ export { default as BackButton } from "./UI/BackButton";
 export { default as RecurringJobsTableBody } from "./Recurring/RecurringJobsTableBody";
 export { default as RecurringJobsTableHead } from "./Recurring/RecurringJobsTableHead";
 export { default as WorkerFilter } from "./Recurring/WorkerFilter";
-export { default as WorkerJobsList } from "./Jobs/Worker/WorkerJobsList";
-export { default as WorkerLogTimeForm } from "./Jobs/Worker/WorkerLogTimeForm";
-export { default as WorkerJobInfo } from "./Jobs/Worker/WorkerJobInfo";
+export { default as WorkerJobsList } from "./WorkerJobs/WorkerJobsList";
+export { default as WorkerLogTimeForm } from "./WorkerJobs/WorkerLogTimeForm";
+export { default as WorkerJobInfo } from "./WorkerJobs/WorkerJobInfo";
 export { default as CreateCustomerReport } from "./Reports/CreateCustomerReport";
 export { default as CreateWorkerReport } from "./Reports/CreateWorkerReport";
 export { default as HomepageText } from "./Homepage/HomepageText";
